Separando a parte do código que executa como root da parte que executa como usuário.
