+++ Instalar.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+python3 InstalarRoot.py
+python3 InstalarUsuario.py
